BinanceAgent/requirements.txt
python-dotenv==1.0.1
requests==2.31.0

# Google AI
google-generativeai==0.8.3

# LangChain
langchain==0.3.14
langchain-core==0.3.25
langchain-google-genai==2.0.6

# FastAPI
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2

